feat: Footer (#51)
* refactor: removes commented style and tidies font setup

* feat: adds footer text

* feat: adds icons to footer and modularises them for reuse with navbar

* feat: adds nav to footer and mdularises for sharing

* feat: adds layour styling and improves nav spacing setup

* feat: adds dynamic copyright year
