+++ styles/Footer.module.sass
@@ -1,3 +1,23 @@
 .footer
-    background-color: coral
     padding: 1rem
+    padding: 0 2rem 1rem
+
+    hr
+        margin-bottom: 1rem
+
+.footerContent
+    display: flex
+    justify-content: space-between
+
+.footerTitle
+    h4
+        margin: 0
+
+.footerNav
+    display: flex
+    flex-direction: column
+    justify-content: space-between
+    text-align: right
+
+    p
+        margin: 0
