+++ styles/Nav.module.sass
@@ -1,0 +1,12 @@
+
+.nav
+    a:not(:first-child)
+        margin-left: 1rem
+
+.navItem,
+.selectedNavItem
+    cursor: pointer
+
+// TODO: choose button styling
+.selectedNavItem
+    color: blue
