+++ styles/Navbar.module.sass
@@ -3,18 +3,7 @@
     flex-direction: row
     justify-content: space-between
     align-items: center
-    // background-color: hotpink
     padding: 1rem 2rem
-
-.navItem,
-.selectedNavItem
-    margin: 0 0.5rem
-    cursor: pointer
-
-// TODO: choose button styling
-.selectedNavItem
-    color: blue
-
 
 .title h3,
 .title h4
@@ -22,7 +11,3 @@
 
 .jobDescription
     font-weight: 300
-
-.icons svg
-    font-size: 1.4rem
-    margin-left: 0.5rem
