+++ styles/SocialIcons.module.sass
@@ -1,0 +1,5 @@
+.icons
+    font-size: 1.4rem
+    
+    a:not(:first-child)
+        margin-left: 0.5rem
