+++ styles/globals.sass
@@ -16,6 +16,6 @@ a
         color-scheme: dark
 
     body
-        color: white
-        background: black
+        color: #FDFDFD
+        background: #1F2028
